ISSUE #1561: Move raw activity data processing to separate pipeline
