+++ tests/Application/Import/ProcessRawActivityData/ProcessRawActivityDataCommandHandlerTest.php
@@ -1,23 +1,32 @@
 <?php
 
-namespace App\Tests\Application\Import\ImportActivitySplits;
+namespace App\Tests\Application\Import\ProcessRawActivityData;
 
-use App\Application\Import\ImportActivitySplits\ImportActivitySplits;
+use App\Application\Import\ProcessRawActivityData\ProcessRawActivityData;
 use App\Domain\Activity\ActivityId;
 use App\Domain\Activity\ActivityWithRawData;
 use App\Domain\Activity\ActivityWithRawDataRepository;
+use App\Domain\Activity\Lap\ActivityLapRepository;
 use App\Domain\Activity\Split\ActivitySplitRepository;
+use App\Domain\Segment\SegmentEffort\SegmentEffortId;
+use App\Domain\Segment\SegmentEffort\SegmentEffortRepository;
+use App\Domain\Segment\SegmentId;
+use App\Domain\Segment\SegmentRepository;
 use App\Infrastructure\CQRS\Command\Bus\CommandBus;
 use App\Infrastructure\Serialization\Json;
+use App\Infrastructure\ValueObject\Measurement\Length\Kilometer;
 use App\Infrastructure\ValueObject\Measurement\UnitSystem;
 use App\Infrastructure\ValueObject\Time\SerializableDateTime;
 use App\Tests\ContainerTestCase;
 use App\Tests\Domain\Activity\ActivityBuilder;
+use App\Tests\Domain\Activity\Lap\ActivityLapBuilder;
 use App\Tests\Domain\Activity\Split\ActivitySplitBuilder;
+use App\Tests\Domain\Segment\SegmentBuilder;
+use App\Tests\Domain\Segment\SegmentEffort\SegmentEffortBuilder;
 use App\Tests\SpyOutput;
 use Spatie\Snapshots\MatchesSnapshots;
 
-class ImportActivitySplitsCommandHandlerTest extends ContainerTestCase
+class ProcessRawActivityDataCommandHandlerTest extends ContainerTestCase
 {
     use MatchesSnapshots;
 
@@ -27,6 +36,36 @@ class ImportActivitySplitsCommandHandlerTest extends ContainerTestCase
     {
         $output = new SpyOutput();
 
+        // Activities for laps.
+        $this->getContainer()->get(ActivityWithRawDataRepository::class)->add(ActivityWithRawData::fromState(
+            ActivityBuilder::fromDefaults()
+                ->withActivityId(ActivityId::fromUnprefixed('with-laps-not-imported'))
+                ->withStartDateTime(SerializableDateTime::fromString('2025-01-01'))
+                ->build(),
+            Json::decode('{"laps":[{"id":52014888132,"resource_state":2,"name":"Lap 1","activity":{"id":14630387617,"visibility":"everyone","resource_state":1},"athlete":{"id":62214940,"resource_state":1},"elapsed_time":643,"moving_time":644,"start_date":"2025-05-29T10:15:42Z","start_date_local":"2025-05-29T12:15:42Z","distance":5010.93,"average_speed":7.78,"max_speed":19.74,"lap_index":1,"split":1,"start_index":0,"end_index":607,"total_elevation_gain":68.2,"average_cadence":87.8,"device_watts":true,"average_watts":175.6,"average_heartrate":162.2,"max_heartrate":171},{"id":52014888145,"resource_state":2,"name":"Lap 2","activity":{"id":14630387617,"visibility":"everyone","resource_state":1},"athlete":{"id":62214940,"resource_state":1},"elapsed_time":6076,"moving_time":6077,"start_date":"2025-05-29T10:26:26Z","start_date_local":"2025-05-29T12:26:26Z","distance":45126.6,"average_speed":7.43,"max_speed":19.88,"lap_index":2,"split":2,"start_index":608,"end_index":6672,"total_elevation_gain":828.8,"average_cadence":87.6,"device_watts":true,"average_watts":179.5,"average_heartrate":164.6,"max_heartrate":176}]}'),
+        ));
+        $this->getContainer()->get(ActivityWithRawDataRepository::class)->add(ActivityWithRawData::fromState(
+            ActivityBuilder::fromDefaults()
+                ->withActivityId(ActivityId::fromUnprefixed('with-laps-but-already-imported'))
+                ->withStartDateTime(SerializableDateTime::fromString('2025-01-10'))
+                ->build(),
+            Json::decode('{"laps":[{"id":52014888132,"resource_state":2,"name":"Lap 1","activity":{"id":14630387617,"visibility":"everyone","resource_state":1},"athlete":{"id":62214940,"resource_state":1},"elapsed_time":643,"moving_time":644,"start_date":"2025-05-29T10:15:42Z","start_date_local":"2025-05-29T12:15:42Z","distance":5010.93,"average_speed":7.78,"max_speed":19.74,"lap_index":1,"split":1,"start_index":0,"end_index":607,"total_elevation_gain":68.2,"average_cadence":87.8,"device_watts":true,"average_watts":175.6,"average_heartrate":162.2,"max_heartrate":171},{"id":52014888145,"resource_state":2,"name":"Lap 2","activity":{"id":14630387617,"visibility":"everyone","resource_state":1},"athlete":{"id":62214940,"resource_state":1},"elapsed_time":6076,"moving_time":6077,"start_date":"2025-05-29T10:26:26Z","start_date_local":"2025-05-29T12:26:26Z","distance":45126.6,"average_speed":7.43,"max_speed":19.88,"lap_index":2,"split":2,"start_index":608,"end_index":6672,"total_elevation_gain":828.8,"average_cadence":87.6,"device_watts":true,"average_watts":179.5,"average_heartrate":164.6,"max_heartrate":176}]}'),
+        ));
+        $this->getContainer()->get(ActivityLapRepository::class)->add(
+            ActivityLapBuilder::fromDefaults()
+                ->withActivityId(ActivityId::fromUnprefixed('with-laps-but-already-imported'))
+                ->withLapNumber(3)
+                ->build()
+        );
+        $this->getContainer()->get(ActivityWithRawDataRepository::class)->add(ActivityWithRawData::fromState(
+            ActivityBuilder::fromDefaults()
+                ->withActivityId(ActivityId::fromUnprefixed('without-laps'))
+                ->withStartDateTime(SerializableDateTime::fromString('2025-01-09'))
+                ->build(),
+            []
+        ));
+
+        // Activities for splits.
         $this->getContainer()->get(ActivityWithRawDataRepository::class)->add(ActivityWithRawData::fromState(
             ActivityBuilder::fromDefaults()
                 ->withActivityId(ActivityId::fromUnprefixed('with-splits-not-imported'))
@@ -56,8 +95,119 @@ class ImportActivitySplitsCommandHandlerTest extends ContainerTestCase
             []
         ));
 
-        $this->commandBus->dispatch(new ImportActivitySplits($output));
+        // Activities for segments.
+        $this->getContainer()->get(ActivityWithRawDataRepository::class)->add(ActivityWithRawData::fromState(
+            ActivityBuilder::fromDefaults()
+                ->withActivityId(ActivityId::fromUnprefixed(1))
+                ->withName('Test activity 1')
+                ->withDeviceName('Zwift')
+                ->withStartDateTime(SerializableDateTime::fromString('2025-01-01'))
+                ->build(),
+            [
+                'segment_efforts' => [
+                    [
+                        'id' => '1',
+                        'start_date_local' => '2023-07-29T09:34:03Z',
+                        'name' => 'Segment Effort One',
+                        'elapsed_time' => 300,
+                        'segment' => [
+                            'id' => '1',
+                            'name' => 'Segment One',
+                            'maximum_grade' => 5.3,
+                            'distance' => 1000,
+                        ],
+                    ],
+                ],
+            ],
+        ));
+        $this->getContainer()->get(ActivityWithRawDataRepository::class)->add(ActivityWithRawData::fromState(
+            ActivityBuilder::fromDefaults()
+                ->withActivityId(ActivityId::fromUnprefixed(2))
+                ->withName('Test activity 2')
+                ->withStartDateTime(SerializableDateTime::fromString('2024-01-01'))
+                ->build(),
+            [
+                'segment_efforts' => [
+                    [
+                        'id' => '2',
+                        'start_date_local' => '2023-07-29T09:34:03Z',
+                        'name' => 'Segment Effort Two',
+                        'segment' => [
+                            'id' => '1',
+                            'name' => 'Segment One',
+                            'maximum_grade' => 5.3,
+                            'distance' => 1000,
+                        ],
+                    ],
+                ],
+            ],
+        ));
+        $this->getContainer()->get(ActivityWithRawDataRepository::class)->add(ActivityWithRawData::fromState(
+            ActivityBuilder::fromDefaults()
+                ->withActivityId(ActivityId::fromUnprefixed(3))
+                ->withName('Test activity 3')
+                ->build(),
+            []
+        ));
+        $this->getContainer()->get(ActivityWithRawDataRepository::class)->add(ActivityWithRawData::fromState(
+            ActivityBuilder::fromDefaults()
+                ->withActivityId(ActivityId::fromUnprefixed(4))
+                ->withName('Test activity 4')
+                ->withStartDateTime(SerializableDateTime::fromString('2024-01-01'))
+                ->build(),
+            [
+                'segment_efforts' => [
+                    [
+                        'id' => '3',
+                        'start_date_local' => '2023-07-29T09:34:03Z',
+                        'name' => 'Segment Effort Two',
+                        'segment' => [
+                            'id' => '2',
+                            'name' => 'Segment Two',
+                            'maximum_grade' => 5.3,
+                            'distance' => 1000,
+                            'starred' => true,
+                        ],
+                    ],
+                ],
+            ],
+        ));
+        $this->getContainer()->get(SegmentEffortRepository::class)->add(
+            SegmentEffortBuilder::fromDefaults()
+                ->withSegmentEffortId(SegmentEffortId::fromUnprefixed(2))
+                ->withSegmentId(SegmentId::fromUnprefixed('1'))
+                ->withActivityId(ActivityId::fromUnprefixed(9542782314))
+                ->withElapsedTimeInSeconds(9.3)
+                ->withAverageWatts(200)
+                ->withDistance(Kilometer::from(0.1))
+                ->build()
+        );
+        $this->getContainer()->get(SegmentEffortRepository::class)->add(
+            SegmentEffortBuilder::fromDefaults()
+                ->withSegmentEffortId(SegmentEffortId::fromUnprefixed(3))
+                ->withSegmentId(SegmentId::fromUnprefixed('2'))
+                ->withActivityId(ActivityId::fromUnprefixed(9542782314))
+                ->withElapsedTimeInSeconds(9.3)
+                ->withAverageWatts(200)
+                ->withDistance(Kilometer::from(0.1))
+                ->build()
+        );
+        $this->getContainer()->get(SegmentRepository::class)->add(
+            SegmentBuilder::fromDefaults()
+                ->withSegmentId(SegmentId::fromUnprefixed('2'))
+                ->withIsFavourite(false)
+                ->build()
+        );
+
+        $this->commandBus->dispatch(new ProcessRawActivityData($output));
         $this->assertMatchesTextSnapshot($output);
+
+        $this->assertMatchesJsonSnapshot(
+            $this->getConnection()->executeQuery('SELECT * FROM Segment')->fetchAllAssociative()
+        );
+        $this->assertMatchesJsonSnapshot(
+            $this->getConnection()->executeQuery('SELECT * FROM SegmentEffort')->fetchAllAssociative()
+        );
     }
 
     #[\Override]
